// Atspm/WebUI/src/api/speedManagement/aTSPMSpeedManagementApi.schemas.ts
export type AggClassification = typeof AggClassification[keyof typeof AggClassification];


export const AggClassification = {
  Total: 'Total',
  Weekend: 'Weekend',
  Weekday: 'Weekday',
} as const;

export interface DoubleDataPoint {
  timestamp: string;
  value: number;
}

export interface Int64DataPoint {
  timestamp: string;
  value: number;
}

export interface AverageAndEightyFifthSeriesData {
  average: DoubleDataPoint[];
  eightyFifth: Int64DataPoint[];
}

export interface SpeedDataDto {
  date: string;
  series: AverageAndEightyFifthSeriesData;
}

export interface CongestionTrackingDto {
  segmentId: string;
  /** @minLength 1 */
  segmentName: string;
  startDate: string;
  endDate: string;
  startingMilePoint: number;
  endingMilePoint: number;
  speedLimit: number;
  data: SpeedDataDto[];
}

export interface CongestionTrackingOptions {
  startDate: string;
  endDate: string;
  sourceId: number[];
  segmentId: string;
}

export interface Coordinate {
  x: number;
  y: number;
  z: number;
  m: number;
  coordinateValue: Coordinate;
  readonly isValid: boolean;
}
// ...
export interface DailySpeedViolationsDto {
  date: string;
  dailyFlow: number;
  dailyViolationsCount: number;
  dailyExtremeViolationsCount: number;
  dailyPercentViolations: number;
  dailyPercentExtremeViolations: number;
}
// ...
export type Dimension = typeof Dimension[keyof typeof Dimension];


export const Dimension = {
  Point: 'Point',
  Curve: 'Curve',
  Surface: 'Surface',
  Collapse: 'Collapse',
  Dontcare: 'Dontcare',
  True: 'True',
  False: 'False',
} as const;

export interface TimeSegmentEffectiveness {
  startDate: string;
  endDate: string;
  flow: number;
  minSpeed: number;
  maxSpeed: number;
  averageSpeed: number;
  /** @nullable */
  averageEightyFifthSpeed?: number | null;
  variability: number;
  percentViolations: number;
  percentExtremeViolations: number;
}

export interface EffectivenessOfStrategiesDto {
  segmentId: string;
  segmentName: string;
  changeInAverageSpeed: number;
  /** @nullable */
  changeInEightyFifthPercentileSpeed?: number | null;
  changeInVariability: number;
  changeInPercentViolations: number;
  changeInPercentExtremeViolations: number;
  speedLimit: number;
  weeklyEffectiveness: TimeSegmentEffectiveness[];
  before: TimeSegmentEffectiveness;
  after: TimeSegmentEffectiveness;
}

export interface EffectivenessOfStrategiesOptions {
  strategyImplementedDate: string;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  endTime?: string | null;
  segmentIds: string[];
}

export interface Envelope {
  readonly isNull: boolean;
  readonly width: number;
  readonly height: number;
  readonly diameter: number;
  readonly minX: number;
  readonly maxX: number;
  readonly minY: number;
  readonly maxY: number;
  readonly area: number;
  readonly minExtent: number;
  readonly maxExtent: number;
  centre: Coordinate;
}

export interface ElevationModel {
  extent: Envelope;
}

export interface EntityRequestOptions {
  coordinates: number[][];
  startDate: string;
  endDate: string;
  /** @nullable */
  directions?: string[] | null;
  /** @nullable */
  sourceId?: number | null;
}

export type TimePeriodFilter = typeof TimePeriodFilter[keyof typeof TimePeriodFilter];


export const TimePeriodFilter = {
  AllDay: 'AllDay',
  OffPeak: 'OffPeak',
  AmPeak: 'AmPeak',
  PmPeak: 'PmPeak',
  MidDay: 'MidDay',
  Evening: 'Evening',
  EarlyMorning: 'EarlyMorning',
} as const;

export type SpeedDataType = typeof SpeedDataType[keyof typeof SpeedDataType];


export const SpeedDataType = {
  H: 'H',
  M: 'M',
} as const;

export type ReportManagementType = typeof ReportManagementType[keyof typeof ReportManagementType];


export const ReportManagementType = {
  Engineer: 'Engineer',
  Leadership: 'Leadership',
} as const;

export interface ExportableReportOptions {
  startDate: string;
  endDate: string;
  timePeriod: TimePeriodFilter;
  aggClassification: AggClassification;
  speedDataType: SpeedDataType;
  reportManagementType: ReportManagementType;
  excludeSourceId: boolean;
  /** @nullable */
  sourceId?: number[] | null;
  excludeRegion: boolean;
  /** @nullable */
  region?: string[] | null;
  excludeFunctionalType: boolean;
  /** @nullable */
  functionalType?: string[] | null;
  excludeAccessCategory: boolean;
  /** @nullable */
  accessCategory?: string[] | null;
  excludeCity: boolean;
  /** @nullable */
  city?: string[] | null;
  excludeCounty: boolean;
  /** @nullable */
  county?: string[] | null;
  /** @nullable */
  order?: string | null;
  /** @nullable */
  limit?: number | null;
}

export type ExportableReportType = typeof ExportableReportType[keyof typeof ExportableReportType];


export const ExportableReportType = {
  ComplianceToSpeedLimit: 'ComplianceToSpeedLimit',
  SpeedVariability: 'SpeedVariability',
  Violations: 'Violations',
  MostImproved: 'MostImproved',
  EffectivenessOfStrategies: 'EffectivenessOfStrategies',
} as const;

export type PrecisionModels = typeof PrecisionModels[keyof typeof PrecisionModels];


export const PrecisionModels = {
  Floating: 'Floating',
  FloatingSingle: 'FloatingSingle',
  Fixed: 'Fixed',
} as const;

export interface PrecisionModel {
  readonly isFloating: boolean;
  readonly maximumSignificantDigits: number;
  scale: number;
  readonly gridSize: number;
  precisionModelType: PrecisionModels;
}
// ...
export interface GeometryFactory {
  precisionModel: PrecisionModel;
  coordinateSequenceFactory: CoordinateSequenceFactory;
  readonly srid: number;
  elevationModel: ElevationModel;
  geometryServices: NtsGeometryServices;
}

export type OgcGeometryType = typeof OgcGeometryType[keyof typeof OgcGeometryType];


export const OgcGeometryType = {
  Point: 'Point',
  LineString: 'LineString',
  Polygon: 'Polygon',
  MultiPoint: 'MultiPoint',
  MultiLineString: 'MultiLineString',
  MultiPolygon: 'MultiPolygon',
  GeometryCollection: 'GeometryCollection',
  CircularString: 'CircularString',
  CompoundCurve: 'CompoundCurve',
  CurvePolygon: 'CurvePolygon',
  MultiCurve: 'MultiCurve',
  MultiSurface: 'MultiSurface',
  Curve: 'Curve',
  Surface: 'Surface',
  PolyhedralSurface: 'PolyhedralSurface',
  TIN: 'TIN',
} as const;

export interface Point {
  factory: GeometryFactory;
  userData?: unknown | null;
  srid: number;
  precisionModel: PrecisionModel;
  readonly numGeometries: number;
  readonly isSimple: boolean;
  readonly isValid: boolean;
  readonly area: number;
  readonly length: number;
  centroid: Point;
  interiorPoint: Point;
  pointOnSurface: Point;
  envelope: Geometry;
  envelopeInternal: Envelope;
  readonly isRectangle: boolean;
  coordinateSequence: CoordinateSequence;
  /** @nullable */
  readonly coordinates?: readonly Coordinate[] | null;
  readonly numPoints: number;
  readonly isEmpty: boolean;
  dimension: Dimension;
  boundaryDimension: Dimension;
  x: number;
  y: number;
  coordinate: Coordinate;
  /** @nullable */
  readonly geometryType?: string | null;
  ogcGeometryType: OgcGeometryType;
  boundary: Geometry;
  z: number;
  m: number;
}

export interface Geometry {
  factory: GeometryFactory;
  userData?: unknown | null;
  srid: number;
  /** @nullable */
  readonly geometryType?: string | null;
  ogcGeometryType: OgcGeometryType;
  precisionModel: PrecisionModel;
  coordinate: Coordinate;
  /** @nullable */
  readonly coordinates?: readonly Coordinate[] | null;
  readonly numPoints: number;
  readonly numGeometries: number;
  readonly isSimple: boolean;
  readonly isValid: boolean;
  readonly isEmpty: boolean;
  readonly area: number;
  readonly length: number;
  centroid: Point;
  interiorPoint: Point;
  pointOnSurface: Point;
  dimension: Dimension;
  boundary: Geometry;
  boundaryDimension: Dimension;
  envelope: Geometry;
  envelopeInternal: Envelope;
  readonly isRectangle: boolean;
}

export interface RouteSpeed {
  createdDate: string;
  binStartTime: string;
  segmentId: string;
  sourceId: number;
  name: string;
  shape: Geometry;
  speedLimit: number;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  county?: string | null;
  /** @nullable */
  averageSpeed?: number | null;
  /** @nullable */
  averageEightyFifthSpeed?: number | null;
  /** @nullable */
  violations?: number | null;
  /** @nullable */
  extremeViolations?: number | null;
  /** @nullable */
  flow?: number | null;
  /** @nullable */
  minSpeed?: number | null;
  /** @nullable */
  maxSpeed?: number | null;
  /** @nullable */
  variability?: number | null;
  /** @nullable */
  percentViolations?: number | null;
  /** @nullable */
  percentExtremeViolations?: number | null;
  /** @nullable */
  avgSpeedVsSpeedLimit?: number | null;
  /** @nullable */
  eightyFifthSpeedVsSpeedLimit?: number | null;
  /** @nullable */
  percentObserved?: number | null;
  hasData: boolean;
}

export interface ImpactType {
  /** @nullable */
  id?: string | null;
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface ImpactDto {
  /** @nullable */
  id?: string | null;
  /** @nullable */
  description?: string | null;
  start: string;
  /** @nullable */
  end?: string | null;
  startMile: number;
  endMile: number;
  speedLimit: number;
  /** @nullable */
  createdOn?: string | null;
  /** @nullable */
  createdBy?: string | null;
  /** @nullable */
  updatedOn?: string | null;
  /** @nullable */
  updatedBy?: string | null;
  /** @nullable */
  deletedOn?: string | null;
  /** @nullable */
  deletedBy?: string | null;
  /** @nullable */
  changeInAverageSpeed?: number | null;
  /** @nullable */
  changeInEightyFifthPercentileSpeed?: number | null;
  /** @nullable */
  changeInVariability?: number | null;
  /** @nullable */
  changeInPercentViolations?: number | null;
  /** @nullable */
  changeInPercentExtremeViolations?: number | null;
  /** @nullable */
  beforeFlow?: number | null;
  /** @nullable */
  beforeMinSpeed?: number | null;
  /** @nullable */
  beforeMaxSpeed?: number | null;
  /** @nullable */
  beforeAverageSpeed?: number | null;
  /** @nullable */
  beforeAverageEightyFifthSpeed?: number | null;
  /** @nullable */
  beforeVariability?: number | null;
  /** @nullable */
  beforePercentViolations?: number | null;
  /** @nullable */
  beforePercentExtremeViolations?: number | null;
  /** @nullable */
  afterFlow?: number | null;
  /** @nullable */
  afterMinSpeed?: number | null;
  /** @nullable */
  afterMaxSpeed?: number | null;
  /** @nullable */
  afterAverageSpeed?: number | null;
  /** @nullable */
  afterAverageEightyFifthSpeed?: number | null;
  /** @nullable */
  afterVariability?: number | null;
  /** @nullable */
  afterPercentViolations?: number | null;
  /** @nullable */
  afterPercentExtremeViolations?: number | null;
  /** @nullable */
  segmentIds?: string[] | null;
  /** @nullable */
  impactTypeIds?: string[] | null;
  /** @nullable */
  impactTypes?: ImpactType[] | null;
  /** @nullable */
  shapes?: Geometry[] | null;
}

export interface SpeedViolationsDto {
  segmentId: string;
  segmentName: string;
  totalFlow: number;
  totalViolationsCount: number;
  totalExtremeViolationsCount: number;
  percentViolations: number;
  percentExtremeViolations: number;
  speedLimit: number;
  dailySpeedViolationsDto: DailySpeedViolationsDto[];
}

export interface SpeedComplianceDto {
  segmentId: string;
  segmentName: string;
  startDate: string;
  endDate: string;
  startingMilePoint: number;
  endingMilePoint: number;
  speedLimit: number;
  average: number;
  avgVsBaseSpeed: number;
  eightyFifth: number;
  eightyFifthPercentileVsBaseSpeed: number;
}

export interface SpeedVariabilityDataDto {
  date: string;
  /** @nullable */
  minSpeed?: number | null;
  /** @nullable */
  maxSpeed?: number | null;
  avgSpeed: number;
  /** @nullable */
  speedVariability?: number | null;
}

export interface SpeedVariabilityDto {
  segmentId: string;
  segmentName: string;
  startDate: string;
  endDate: string;
  startingMilePoint: number;
  endingMilePoint: number;
  speedLimit: number;
  data: SpeedVariabilityDataDto[];
}

export interface ExportableReportResult {
  exportableReportType: ExportableReportType;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  reportData?: RouteSpeed[] | null;
  /** @nullable */
  impacts?: ImpactDto[] | null;
  /** @nullable */
  congestionTrackingCharts?: CongestionTrackingDto[] | null;
  /** @nullable */
  speedViolationCharts?: SpeedViolationsDto[] | null;
  /** @nullable */
  speedComplianceCharts?: SpeedComplianceDto[] | null;
  /** @nullable */
  speedVariabilityCharts?: SpeedVariabilityDto[] | null;
  /** @nullable */
  effectivenessOfStrategiesCharts?: EffectivenessOfStrategiesDto[] | null;
}